refactor most config into separate components
